requirements.txt
Django==1.10
requests==2.11.1
line-bot-sdk==1.0.0
beautifulsoup4==4.5.1
gunicorn==19.0.0

